+++ HLSLDecompiler/DecompileHLSL.h
@@ -4,11 +4,14 @@
 #include <vector>
 #include <string>
 
+// The goal with this file is too keep it as much C++ as possible.
+// So we avoid adding Windows specific stuff if at all possible.
+
 struct ParseParameters
 {
 	const void *bytecode;
 	const char *decompiled;
-	long decompiledSize;
+	size_t decompiledSize;
 
 	bool fixSvPosition;
 	bool fixLightPosition;
